fix loop request listComments
