+++ src/contexts/CommentsContext.tsx
@@ -1,5 +1,5 @@
 import { useDisclosure, useToast } from "@chakra-ui/react";
-import { ReactNode, createContext } from "react";
+import { ReactNode, createContext, useEffect, useState } from "react";
 import api from "../services/api";
 import { iCreateComment } from "../components/CardComment/types";
 
@@ -11,13 +11,38 @@ interface CommentContextValues {
   registerComment(data: iComment): void;
   deleteComment(commentId: number): void;
   updateComment(data: iCreateComment, idComment: number): void;
+  comments: iComment[] | undefined
 }
 
 interface iComment {
-  id?: number;
-  comment?: string;
-  user_id?: number;
-  car_id?: number;
+  id: number;
+  comment: string;
+  createdAt: string;
+  user_id: {
+    id: number;
+    name: string;
+    email: string;
+    cpf: string;
+    phone: string;
+    birthDate: string;
+    description: string;
+    admin: boolean;
+    password: string;
+    reset_token: null;
+  };
+  car_id: {
+    id: number;
+    brand: string;
+    model: string;
+    year: number;
+    fuel: number;
+    km: number;
+    color: string;
+    fipePrice: number;
+    sellPrice: number;
+    description: string;
+    isActive: true;
+  };
 }
 
 export const CommentContext = createContext({} as CommentContextValues);
@@ -28,6 +53,35 @@ export const CommentProvider = ({ children }: CommentsProvider) => {
   const token = localStorage.getItem("motors-shop:token");
 
   const { isOpen, onOpen, onClose } = useDisclosure()
+
+  const [comments, setComments] = useState<iComment[] | undefined>([]);
+
+  const userIdString = localStorage.getItem("motors-shop:user")
+
+  const product = localStorage.getItem("id-product-page:");
+
+  const userId = userIdString ? JSON.parse(userIdString) : null
+
+  const listComments = async () => {
+    if (product) {
+      const idProduct = JSON.parse(product);
+
+      try {
+        const response = await api.get(`/cars/${idProduct.id}/comments`);
+        setComments(response.data);
+        console.log(comments);
+        
+        
+      } catch (error) {
+        console.error(error);
+      }
+    }
+  };
+
+  useEffect(() => {
+
+    listComments();
+  }, []);
 
   const registerComment = async (data: iComment) => {
     const product = localStorage.getItem("id-product-page:");
@@ -57,6 +111,8 @@ export const CommentProvider = ({ children }: CommentsProvider) => {
             status: "success",
             isClosable: true,
           });
+
+          listComments()
         } catch (error) {
           toast({
             title: "Ops, algo deu errado :(",
@@ -82,6 +138,9 @@ export const CommentProvider = ({ children }: CommentsProvider) => {
         status: "success",
         isClosable: true,
       });
+
+      listComments()
+
     } catch (error) {
       toast({
         title: "Ops, algo deu errado :(",
@@ -100,6 +159,8 @@ export const CommentProvider = ({ children }: CommentsProvider) => {
         },
         timeout: 10000,
       })
+
+      listComments()
 
       onClose()
 
@@ -120,7 +181,7 @@ export const CommentProvider = ({ children }: CommentsProvider) => {
   }
 
   return (
-    <CommentContext.Provider value={{ registerComment, deleteComment, updateComment }}>
+    <CommentContext.Provider value={{ registerComment, deleteComment, updateComment, comments }}>
       {children}
     </CommentContext.Provider>
   );
